+++ pkg/providers/vsphere/upgrade/virtualmachine/backfill/suite_test.go
@@ -1,0 +1,18 @@
+// Copyright (c) 2026 Broadcom. All Rights Reserved.
+// Broadcom Confidential. The term "Broadcom" refers to Broadcom Inc.
+// and/or its subsidiaries.
+// SPDX-License-Identifier: Apache-2.0
+
+package backfill_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestBackfill(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Schema Upgrade Backfill Suite")
+}
