+++ pkg/providers/vsphere/virtualmachine/virtualmachine_suite_test.go
@@ -23,9 +23,6 @@ func vcSimTests() {
 	Describe("Snapshot", Label(testlabels.VCSim), snapShotTests)
 	Describe("ExtraConfig", Label(testlabels.VCSim), extraConfigTests)
 	Describe("CleanupOnDelete", Label(testlabels.VCSim), cleanupOnDeleteTests)
-	Describe("NetworkInterfaceTypes", networkInterfaceTypeTests)
-	Describe("BackfillExtraConfig", backfillExtraConfigTests)
-	Describe("BackfillNICDeviceProperties", backfillNICDevicePropertiesTests)
 }
 
 var suite = builder.NewTestSuite()
